Move calibration model to EnggUtils
